Split VLM safetensor spool
